Rename path_filters.hpp -> path_sources.hpp
